Updated Entities, add column length

+++ Web/BrouwershuisApp/src/com/brouwershuis/db/model/Role.java
@@ -20,24 +20,27 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 @Entity
 @Table(name = "role")
 public class Role implements Serializable{
+	
     /**
 	 * 
 	 */
 	private static final long serialVersionUID = 1L;
-	private Long id;
+	
+	private int id;
+	
     private String name;
+    
     private Set<User> users;
     
-
 	private Date timesTamp;
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
